Removed current output, changed DAC to single channel
Current input on the MOGLabs DLC doesn't allow for a DC bias in
the absence of the fast lock.  So I removed the current output,
and therefore removed the 2 channel DAC.  I have kept the 2 channel
op-amp, and now I have two inverting amplifiers where the gain of the
second stage can be adjusted through a trimpot.